pyproject.toml
[project]
name = "llmd-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "fastapi>=0.120.1",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "ruff>=0.14.2",
    "uvicorn>=0.38.0",
]
